test(libutil): split finder by behaviour family
Split finder.test.js (481 LOC) into finder-discovery.test.js (path
discovery: constructor, withLogger, findUpward, findProjectRoot,
findData, findPackagePath, findGeneratedPath, plus the collaborator-
config injection cases) and finder-symlinks.test.js (filesystem
mutation: createSymlink, createPackageSymlinks). Shared beforeEach/
afterEach tempDir setup reproduced per sibling. Assertions unchanged;
28 tests preserved.

Spec 0640 § C (large files), design Decision 5.
